requirements.txt
requests
bs4
lxml
streamlit
streamlit-chat
python-dotenv
pinecone-client
pinecone-text
langchain
nltk
sentence-transformers
langchain-community
langchain-groq
torch
torchvision
langchain_huggingface
openai
tiktoken
